+++ baseline/include/BloomedSurfingTrie.h
@@ -15,7 +15,7 @@ namespace range_filtering {
         explicit BloomedSurfingTrie(std::vector<std::string> &keys, uint64_t max_suffix_length,
                                     uint32_t BF_size, double max_penalty);
         bool lookupPrefix(const std::string &prefix) override;
-        unsigned long long getMemoryUsage() const override;
+        uint64_t getMemoryUsage() const override;
         std::string getName() const override { return "BloomedSurfingTrie"; }
 
     protected:
@@ -25,7 +25,7 @@ namespace range_filtering {
 
             TrieNode(BloomedSurfingTrie const& trie);
             virtual bool lookupNode(const std::string &key, uint64_t position);
-            virtual unsigned long long getMemoryUsage() const;
+            virtual uint64_t getMemoryUsage() const;
 
         private:
             std::map<char, TrieNode *> children_;
@@ -33,7 +33,7 @@ namespace range_filtering {
             void insertKeys(uint64_t position, std::vector<std::string> &keys);
             void insertChildNode(char current, uint64_t position, std::vector<std::string> &keys);
 
-            unsigned long long getChildrenMemoryUsage() const;
+            uint64_t getChildrenMemoryUsage() const;
 
             friend class BloomedSurfingTrie;
         };
@@ -42,7 +42,7 @@ namespace range_filtering {
         protected:
             LeafNode(BloomedSurfingTrie const& trie, uint8_t suffix);
             bool lookupNode(const std::string &key, uint64_t position) override;
-            unsigned long long getMemoryUsage() const override;
+            uint64_t getMemoryUsage() const override;
 
             uint8_t suffix_;
