+++ baseline/include/PrefixFilter.h
@@ -7,7 +7,7 @@ namespace range_filtering {
 class PrefixFilter {
 public:
     virtual bool lookupPrefix(const std::string& prefix) = 0;
-    virtual unsigned long long getMemoryUsage() const = 0;
+    virtual uint64_t getMemoryUsage() const = 0;
     virtual std::string getName() const = 0;
 };
 }
