+++ baseline/src/RestrainedSplashyTrie.cpp
@@ -7,13 +7,13 @@ namespace range_filtering {
         children_ = std::map<char, TrieNode*>();
     }
 
-    unsigned long long RestrainedSplashyTrie::TrieNode::getMemoryUsage() const {
+    uint64_t RestrainedSplashyTrie::TrieNode::getMemoryUsage() const {
         // char + 64-bit pointer per each child + boolean flag
         uint64_t node_size_bits = (8 + 64) * children_.size() + 1;
         return node_size_bits + getChildrenMemoryUsage();
     }
 
-    unsigned long long RestrainedSplashyTrie::TrieNode::getChildrenMemoryUsage() const {
+    uint64_t RestrainedSplashyTrie::TrieNode::getChildrenMemoryUsage() const {
         uint64_t children_size = 0;
         for (auto child : children_) {
             children_size += child.second->getMemoryUsage();
@@ -21,7 +21,7 @@ namespace range_filtering {
         return children_size;
     }
 
-    unsigned long long RestrainedSplashyTrie::LeafNode::getMemoryUsage() const {
+    uint64_t RestrainedSplashyTrie::LeafNode::getMemoryUsage() const {
         // This is not entirely true in this implementation as no matter what's the size of suffix,
         // I keep 8bits per suffix, but for the sake of comparison let's assume they're more packed...
         return trie_.max_suffix_length_;
@@ -95,7 +95,7 @@ namespace range_filtering {
         return key_suffix == suffix_;
     }
 
-    unsigned long long RestrainedSplashyTrie::getMemoryUsage() const {
+    uint64_t RestrainedSplashyTrie::getMemoryUsage() const {
         auto bytes = (unsigned long long) ((64. + root->getMemoryUsage()) / 8.0);
         return bytes + 1;
     }
